requirements.txt
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.18
Pillow==2.3.0
Werkzeug==0.9.4
argparse==1.2.1
gunicorn==18.0
httplib2==0.8
itsdangerous==0.23
requests==2.2.1
six==1.5.2
twilio==3.6.6
wsgiref==0.1.2
